requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
click>=8.0.0
gitpython>=3.1.0
colorama>=0.4.4
pyyaml>=6.0.0
